requirements.txt
boto3>=1.26.0
watchdog>=2.1.0
PyYAML>=6.0
